c_python/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# YOLOv8 (yolov8n_balanced best.pt)
ultralytics>=8.0.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
